# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Posts",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/websites": {
            "get": {
                "summary": "List of websites ",
                "operationId": "0bb79b19130274c6d481eb30ba51a3e6",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string"
                                        },
                                        "hostname": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "id": 10,
                                    "hostname": "https://kun.uz"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create new Website",
                "operationId": "dd95e31bd8230654ab786863b3b30149",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "hostname": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "hostname": "https://kun.uz"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "hostname": "https://kun.uz"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Website exists",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Validation error.",
                                        "value": {
                                            "error": "website exists"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Server error.",
                                        "value": {
                                            "error": "Internal error: something went wrong, try again please"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/websites/{id}": {
            "get": {
                "summary": "Get single post",
                "operationId": "fd38632e10c3d1b3a05523fb4bb226d8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "website id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "examples": {
                            "int": {
                                "summary": "An int value.",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "id": 1,
                                            "hostname": "https://kun.uz"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "delete": {
                "summary": "Delete single website",
                "operationId": "bc7683a068ff091bcd41996e2abe3ebb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "website id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "examples": {
                            "int": {
                                "summary": "An int value.",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal error",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Server error.",
                                        "value": {
                                            "error": "Internal error: something went wrong, try again please"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}